+++ cmd/src/batch_exec.go
@@ -83,7 +83,6 @@ func executeBatchSpecInWorkspaces(ctx context.Context, ui *ui.JSONLines, opts ex
 	svc := service.New(&service.Opts{
 		AllowUnsupported: opts.flags.allowUnsupported,
 		AllowIgnored:     opts.flags.allowIgnored,
-		AllowFiles:       false,
 		Client:           opts.client,
 	})
 	if err := svc.DetermineFeatureFlags(ctx); err != nil {